requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
